+++ app/nvmf_tgt/nvmf_tgt.c
@@ -46,9 +46,6 @@
 #include "spdk/log.h"
 #include "spdk/nvme.h"
 
-#define SPDK_NVMF_BUILD_ETC "/usr/local/etc/nvmf"
-#define SPDK_NVMF_DEFAULT_CONFIG SPDK_NVMF_BUILD_ETC "/nvmf.conf"
-
 static struct spdk_poller *g_acceptor_poller = NULL;
 
 static TAILQ_HEAD(, nvmf_tgt_subsystem) g_subsystems = TAILQ_HEAD_INITIALIZER(g_subsystems);
@@ -270,32 +267,6 @@ nvmf_tgt_shutdown_subsystem_by_nqn(const char *nqn)
 }
 
 static void
-usage(void)
-{
-	struct spdk_app_opts opts;
-
-	spdk_app_opts_init(&opts);
-
-	printf("nvmf [options]\n");
-	printf("options:\n");
-	printf(" -c config  - config file (default %s)\n", SPDK_NVMF_DEFAULT_CONFIG);
-	printf(" -e mask    - tracepoint group mask for spdk trace buffers (default 0x0)\n");
-	printf(" -m mask    - core mask for DPDK\n");
-	printf(" -i shared memory ID (optional)\n");
-	printf(" -l facility - use specific syslog facility (default %s)\n",
-	       opts.log_facility);
-	printf(" -n channel number of memory channels used for DPDK\n");
-	printf(" -p core    master (primary) core for DPDK\n");
-	printf(" -s size    memory size in MB for DPDK\n");
-
-	spdk_tracelog_usage(stdout, "-t");
-
-	printf(" -v         - verbose (enable warnings)\n");
-	printf(" -H         - show this usage\n");
-	printf(" -d         - disable coredump file enabling\n");
-}
-
-static void
 acceptor_poll(void *arg)
 {
 	spdk_nvmf_acceptor_poll();
@@ -361,85 +332,12 @@ This is the main file.
 */
 
 int
-main(int argc, char **argv)
+spdk_nvmf_tgt_start(struct spdk_app_opts *opts)
 {
-	int ch;
 	int rc;
-	struct spdk_app_opts opts = {};
 
-	/* default value in opts */
-	spdk_app_opts_init(&opts);
-
-	opts.name = "nvmf";
-	opts.config_file = SPDK_NVMF_DEFAULT_CONFIG;
-	opts.max_delay_us = 1000; /* 1 ms */
-
-	while ((ch = getopt(argc, argv, "c:de:i:l:m:n:p:qs:t:DH")) != -1) {
-		switch (ch) {
-		case 'd':
-			opts.enable_coredump = false;
-			break;
-		case 'c':
-			opts.config_file = optarg;
-			break;
-		case 'i':
-			opts.shm_id = atoi(optarg);
-			break;
-		case 'l':
-			opts.log_facility = optarg;
-			break;
-		case 't':
-			rc = spdk_log_set_trace_flag(optarg);
-			if (rc < 0) {
-				fprintf(stderr, "unknown flag\n");
-				usage();
-				exit(EXIT_FAILURE);
-			}
-#ifndef DEBUG
-			fprintf(stderr, "%s must be rebuilt with CONFIG_DEBUG=y for -t flag.\n",
-				argv[0]);
-			usage();
-			exit(EXIT_FAILURE);
-#endif
-			break;
-		case 'm':
-			opts.reactor_mask = optarg;
-			break;
-		case 'n':
-			opts.dpdk_mem_channel = atoi(optarg);
-			break;
-		case 'p':
-			opts.dpdk_master_core = atoi(optarg);
-			break;
-		case 's':
-			opts.dpdk_mem_size = atoi(optarg);
-			break;
-		case 'e':
-			opts.tpoint_group_mask = optarg;
-			break;
-		case 'q':
-			spdk_g_notice_stderr_flag = 0;
-			break;
-		case 'D':
-		case 'H':
-		default:
-			usage();
-			exit(EXIT_SUCCESS);
-		}
-	}
-
-	if (spdk_g_notice_stderr_flag == 1 &&
-	    isatty(STDERR_FILENO) &&
-	    !strncmp(ttyname(STDERR_FILENO), "/dev/tty", strlen("/dev/tty"))) {
-		printf("Warning: printing stderr to console terminal without -q option specified.\n");
-		printf("Suggest using -q to disable logging to stderr and monitor syslog, or\n");
-		printf("redirect stderr to a file.\n");
-		printf("(Delaying for 10 seconds...)\n");
-		sleep(10);
-	}
-
-	opts.shutdown_cb = spdk_nvmf_shutdown_cb;
-	spdk_app_init(&opts);
+	opts->shutdown_cb = spdk_nvmf_shutdown_cb;
+	spdk_app_init(opts);
 
 	printf("Total cores available: %d\n", spdk_lcore_count());
 	/* Blocks until the application is exiting */
